Add ability to reconfigure FPGA in running system
Currently this capability is only exposed by the
/sys/devices/platform/omapdss/manager2/panel-notle-dpi/reset
sysfs node.  Writing a non zero value resets and reconfigures the FPGA,
writing zero or a non-numeric value merely resets (behavior for all
values prior to this change).

FPGA image is searched for under the name dss_fpga.img using the
kernel's request_firmware interface which looks in a list of directories
managed by the init process.  These are:

/system/etc/firmware/
/system/vendor/firmware/

Change-Id: I087c999955258f69411841ebc9ee5661f308b490

Conflicts:
	drivers/video/omap2/displays/panel-notle-dpi.c