Move invensense suspend handlers to noirq phase
The invensense suspend / resume handlers use a lock that the irq handler
also uses. With the suspend handler in the normal phase, it is possible
for the interrupt and suspend handler to deadlock with each other.
Moving the suspend handler to noirq guarantees that no interrupts
from the device will be processed between suspend and resume, so
deadlock is avoided.

Change-Id: I950f6cdb8ac28e9311a130a41bbd78e0a414491c