Read gas gauge flags on power supply status request.
The gas gauge flags register is used to determine the charge status.
When power supply core requests this satus, read the most current
flags value instead of using the cached value to ensure an accurate
charge status.

This may remedy an issue where the device UI reports "charging" for
a short time after the device is unplugged.

Change-Id: I931983ffaa0753d0b39a15f11d53a3a8abea8095