Switch sysfs node to indicate to userspace i2c controller timeout
Created a switch node for each of the omap i2c busses.
The switch will flip (state=1) if omaps i2c controller
timed out during data transmisison. This is the kernel
side of the mechanism of sensing i2c bus failures. Its
a workaround for ESD issue. When the device being
shocked a lot of the time i2c.4 goes away and device
is unusable. Before we get a HW fix, request a graceful
shutdown from the userland by android.

Change-Id: Iecc04bb9d6d542db99da9b76104c60b4527d690c