Use I2C_CLIENT_WAKE flag to solve PM issue
According to LKML commentary, the preferred method to solve issues
with the ordering of resuming I2C devices during IRQ is to use the
I2C_CLIENT_WAKE flag. This is supposed to ensure that the I2C
driver resumes before the I2C client bottom-half runs. Also
re-ordered the I2C driver init to occur before the IMU init.

Change-Id: I955f986ae1b0b3d43b43de17ec43607919f47a3c