Implement invalid charger signaling to userspace.
Upon certain charger fault conditions (VBUS > VBAT, anticollapse, poor src),
signal to userspace that the charger is not good via the "invalid_charger"
switchdev node.

Continue to leave charging enabled to make the best effort to charge, in
case the charger fault is temporary. However, only unset the switchdev
node upon vbus removal -- the charge fault interrupt is not reliable
for this purpose.

Change-Id: I932f24869687ae0093cdd64034147713d5394235