Strengthen i2c timeout requirement for switch device pulse
Now the switch device will only pulse if the i2c controller
had 8 timeouts in a row, without any success in between.
This is meant to reduce the probability of a race condition
in suspend/resume cycles to generate erroneous shutdowns.
BatteryService.java in frameworks/base looks at this switch
node and requests the system shutdown, thinking it is a
sign of ESD shock to the system.

Change-Id: If590ba4f5b938815cd5db18655492704076ee263