Fix for potential deadlock in glasshhub.
In glasshub_suspend(), after the mutex is held, there is a possibility that
the glasshub_threaded_irq_handler would run and try to get the same mutex.
A deadlock could be resulted as interrupt has already been disabled during
this phase of device suspension.

A flag is set in glasshub_suspend() after the mutex is held to ensure future
irq handlers will skip the processing. Most importantly, the pursuit of the mutex.

Change-Id: I8b7f28ae9223055d3cd5a721d8307eb730ba75f6