Enable charger driver to determine charger type at early boot.
The normal mechanism for the charger driver to determine the charger
type, and therefor the charger current limit, is through the
twl6030 usb notifier. This notifier is called whenever the charge
detection takes place. Unfortunately, during early boot, the charge
detection takes place before the charger driver can register with
the notifier chain.

Add a call to allow the charger driver to query the charger type
to work around this problem.

Change-Id: I81e3e159805e2c6afca5d566891074fd9dd9d357