Moved glasshub irq bottom half to freezable workqueue.
This has the benefit of ensuring there are no bottom half activities
during device driver suspension.

At glasshub_suspend_noirq(), if there are any pending work items that
need to be processed, the driver will abort the suspend. This ensures
all IRQs from the glasshub will be handled.

Due to the fact that frozen work items are not restored until all device
drivers have been resumed, there is an added delay in the resume path for
wink and don/off notifications to be sent up to the userspace.

Change-Id: Iaf413d49f8b500562ac2eb88724b322671fa2525