Fix possible race condition in glasshub diagnostics
This is a long shot, but there is a race condition that could
occur where the IRQ happens after the flag check in suspend_noirq
but before the suspended flag is set. This patch closes that
hole so that the diagnostic should print if we ever get an interrupt
in the suspend_noirq routine. This won't fix the I2C timeout issue,
but might give us a clue what's happening.

Change-Id: I324a4e10d7d947e04902a815a59c6ca49a3592b1