Add power management refcounting to glasshub driver
The I2C driver power management is not sufficient to protect
against read errors occuring in the IRQ threaded handler due
to the I2C driver entering suspend. This CL adds all the
power management hooks required to keep the glasshub driver
out of suspend in the bottom half of the driver. The root
cause is still unknown, but suspected to be some sort of
timing issue related to a clock being brought down and back
up again in a short period of time.

Change-Id: I7b34538f05c48a9af8a33c99003ab6b28208583b